EE-9764 Moved NameNormalizers into their own package
