+++ src/main/java/uk/gov/digital/ho/pttg/application/HmrcHateoasClient.java
@@ -15,7 +15,7 @@ import uk.gov.digital.ho.pttg.api.RequestHeaderData;
 import uk.gov.digital.ho.pttg.application.domain.Individual;
 import uk.gov.digital.ho.pttg.application.namematching.NameMatchingCandidatesService;
 import uk.gov.digital.ho.pttg.application.namematching.CandidateName;
-import uk.gov.digital.ho.pttg.application.util.NameNormalizer;
+import uk.gov.digital.ho.pttg.application.util.namenormalizer.NameNormalizer;
 import uk.gov.digital.ho.pttg.dto.*;
 import uk.gov.digital.ho.pttg.dto.sasummary.Summary;
 import uk.gov.digital.ho.pttg.dto.sasummary.SummarySelfAssessment;
