+++ src/main/java/uk/gov/digital/ho/pttg/application/SpringConfiguration.java
@@ -26,10 +26,10 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import uk.gov.digital.ho.pttg.api.RequestHeaderData;
 import uk.gov.digital.ho.pttg.application.retry.RetryProperties;
 import uk.gov.digital.ho.pttg.application.retry.RetryTemplateBuilder;
-import uk.gov.digital.ho.pttg.application.util.CompositeNameNormalizer;
-import uk.gov.digital.ho.pttg.application.util.DiacriticNameNormalizer;
-import uk.gov.digital.ho.pttg.application.util.MaxLengthNameNormalizer;
-import uk.gov.digital.ho.pttg.application.util.NameNormalizer;
+import uk.gov.digital.ho.pttg.application.util.namenormalizer.CompositeNameNormalizer;
+import uk.gov.digital.ho.pttg.application.util.namenormalizer.DiacriticNameNormalizer;
+import uk.gov.digital.ho.pttg.application.util.namenormalizer.MaxLengthNameNormalizer;
+import uk.gov.digital.ho.pttg.application.util.namenormalizer.NameNormalizer;
 
 import java.text.SimpleDateFormat;
 import java.time.Clock;
