+++ src/test/java/uk/gov/digital/ho/pttg/application/util/namenormalizer/CompositeNameNormalizerTest.java
@@ -1,9 +1,11 @@
-package uk.gov.digital.ho.pttg.application.util;
+package uk.gov.digital.ho.pttg.application.util.namenormalizer;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
+import uk.gov.digital.ho.pttg.application.util.namenormalizer.CompositeNameNormalizer;
+import uk.gov.digital.ho.pttg.application.util.namenormalizer.NameNormalizer;
 import uk.gov.digital.ho.pttg.dto.HmrcIndividual;
 
 import static org.assertj.core.api.Assertions.assertThat;
