+++ src/test/java/uk/gov/digital/ho/pttg/application/util/namenormalizer/DiacriticNameNormalizerTest.java
@@ -1,8 +1,9 @@
-package uk.gov.digital.ho.pttg.application.util;
+package uk.gov.digital.ho.pttg.application.util.namenormalizer;
 
 import com.fasterxml.jackson.databind.MappingIterator;
 import com.fasterxml.jackson.dataformat.csv.CsvMapper;
 import org.junit.Test;
+import uk.gov.digital.ho.pttg.application.util.namenormalizer.DiacriticNameNormalizer;
 import uk.gov.digital.ho.pttg.dto.HmrcIndividual;
 
 import java.io.File;
